+++ src/GTEMPO.jl
@@ -49,7 +49,7 @@ export branch, correlationfunction
 
 # influence functional
 export InfluenceFunctionalAlgorithm, PartialIF, XTRGIF, ExactTTIIF, TDVPIF, partialif_hybrid, partialif_hybrid_naive
-export influenceoperator, influenceoperatorexponential, differentialinfluencefunctional
+export influenceoperators, influenceoperatorsteppers, influenceoperatorstepper, influencefunctional
 export hybriddynamics, hybriddynamics!, hybriddynamics_naive, hybriddynamics_naive!, hybriddynamicsstepper, hybriddynamicsstepper!
 export retardedinteractdynamics_naive, retardedinteractdynamics_naive!