+++ src/influencefunctional/hybridization/ttiif/realtime.jl
@@ -1,7 +1,7 @@
 const _AllowedRealGrassmannOrdering = Union{A1Ā1a1ā1B1B̄1b1b̄1, A1Ā1B1B̄1a1ā1b1b̄1}
 
 # full influence operator, only works for ordering A1Ā1a1ā1B1B̄1b1b̄1 and A1Ā1B1B̄1a1ā1b1b̄1
-function influenceoperator(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction; 
+function influenceoperators(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction; 
 							band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion())
 	η⁺⁺, η⁺⁻, η⁻⁺, η⁻⁻ = _get_signed_corr(lattice, corr, band)
 	mpoj1, mpoj2, mpoj3, mpoj4 = ti_mpotensor(η⁺⁺, algexpan), ti_mpotensor(η⁺⁻, algexpan), ti_mpotensor(η⁻⁺, algexpan), ti_mpotensor(η⁻⁻, algexpan)
@@ -17,7 +17,7 @@ function influenceoperator(lattice::RealGrassmannLattice{<:_AllowedRealGrassmann
 	return mpo1, mpo2, mpo3, mpo4
 end
 
-function influenceoperatorexponential(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::FirstOrderStepper; 
+function influenceoperatorsteppers(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::FirstOrderStepper; 
 										band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion())
 	η⁺⁺, η⁺⁻, η⁻⁺, η⁻⁻ = _get_signed_corr(lattice, corr, band)
 	mpoj1, mpoj2, mpoj3, mpoj4 = ti_mpotensor(η⁺⁺, algexpan), ti_mpotensor(η⁺⁻, algexpan), ti_mpotensor(η⁻⁺, algexpan), ti_mpotensor(η⁻⁻, algexpan)
@@ -31,7 +31,7 @@ function influenceoperatorexponential(lattice::RealGrassmannLattice{<:_AllowedRe
 	mpo4 = _fit_to_lattice(lattice, h4, _JW, band, :-, :-) 
 	return mpo1, mpo2, mpo3, mpo4
 end
-function influenceoperatorexponential(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::ComplexStepper; 
+function influenceoperatorsteppers(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::ComplexStepper; 
 										band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion())
 	η⁺⁺, η⁺⁻, η⁻⁺, η⁻⁻ = _get_signed_corr(lattice, corr, band)
 	mpoj1, mpoj2, mpoj3, mpoj4 = ti_mpotensor(η⁺⁺, algexpan), ti_mpotensor(η⁺⁻, algexpan), ti_mpotensor(η⁻⁺, algexpan), ti_mpotensor(η⁻⁻, algexpan)
@@ -51,21 +51,21 @@ function influenceoperatorexponential(lattice::RealGrassmannLattice{<:_AllowedRe
 	mpo4a, mpo4b = _fit_to_lattice(lattice, h4a, _JW, band, :-, :-), _fit_to_lattice(lattice, h4b, _JW, band, :-, :-) 
 	return (mpo1a, mpo1b), (mpo2a, mpo2b), (mpo3a, mpo3b), (mpo4a, mpo4b)
 end
-function differentialinfluencefunctional(lattice::RealGrassmannLattice{O}, corr::RealCorrelationFunction, dt::Real, alg::TimeEvoMPOAlgorithm, algmult::DMRGAlgorithm;
+function influenceoperatorstepper(lattice::RealGrassmannLattice{O}, corr::RealCorrelationFunction, dt::Real, alg::TimeEvoMPOAlgorithm, algmult::DMRGAlgorithm;
 										band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion()) where O
 	if !(OrderingStyle(lattice) isa _AllowedRealGrassmannOrdering)
 		lattice2 = similar(lattice, ordering = A1Ā1a1ā1B1B̄1b1b̄1())
-		mps = _differentialinfluencefunctional(lattice2, corr, dt, alg, algmult; band=band, algexpan=algexpan)
+		mps = _influenceoperatorstepper(lattice2, corr, dt, alg, algmult; band=band, algexpan=algexpan)
 		_, mps2 = changeordering(O, lattice2, mps, trunc=algmult.trunc)
 		return mps2
 	else
-		return _differentialinfluencefunctional(lattice, corr, dt, alg, algmult; band=band, algexpan=algexpan)
+		return _influenceoperatorstepper(lattice, corr, dt, alg, algmult; band=band, algexpan=algexpan)
 	end
 end
 
-function _differentialinfluencefunctional(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::FirstOrderStepper, 
+function _influenceoperatorstepper(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::FirstOrderStepper, 
 											algmult::DMRGAlgorithm; band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion())
-	h1, h2, h3, h4 = influenceoperatorexponential(lattice, corr, dt, alg, band=band, algexpan=algexpan)
+	h1, h2, h3, h4 = influenceoperatorsteppers(lattice, corr, dt, alg, band=band, algexpan=algexpan)
 	mps = h1 * vacuumstate(lattice)
 	tmp = h2 * vacuumstate(lattice)
 	mps = mult(mps, tmp, algmult)
@@ -75,9 +75,9 @@ function _differentialinfluencefunctional(lattice::RealGrassmannLattice{<:_Allow
 	mps = mult(mps, tmp, algmult)
 	return mps
 end
-function _differentialinfluencefunctional(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::ComplexStepper, 
+function _influenceoperatorstepper(lattice::RealGrassmannLattice{<:_AllowedRealGrassmannOrdering}, corr::RealCorrelationFunction, dt::Real, alg::ComplexStepper, 
 											algmult::DMRGAlgorithm; band::Int=1, algexpan::ExponentialExpansionAlgorithm=PronyExpansion())
-	(h1a, h1b), (h2a, h2b), (h3a, h3b), (h4a, h4b) = influenceoperatorexponential(lattice, corr, dt, alg, band=band, algexpan=algexpan)
+	(h1a, h1b), (h2a, h2b), (h3a, h3b), (h4a, h4b) = influenceoperatorsteppers(lattice, corr, dt, alg, band=band, algexpan=algexpan)
 	mps1 = h1a * vacuumstate(lattice)
 	tmp = h1b * vacuumstate(lattice)
 	mps = mult(mps1, tmp, algmult)